Extract testHelpers.cpp, add getTestInputFilePath() and getTestOutputFilePath()
