+++ src/libloaders_test/libdataxp_test.h
@@ -4,6 +4,7 @@
 //
 #pragma once
 
+#include <string>
 #include <fstream>
 #include <sstream>
 #include <vector>
@@ -20,6 +21,8 @@ stringstream makeAptDat(const vector<string>& lines);
 shared_ptr<HostServices> makeHost();
 shared_ptr<ControlledAirspace> makeAirspace(double centerLat, double centerLon, float radiusNm, const string& name);
 void openTestInputStream(const string& fileName, ifstream& str);
+string getTestInputFilePath(const string& fileName);
+string getTestOutputFilePath(const string& fileName);
 void assertRunwaysExist(shared_ptr<Airport> airport, const vector<string>& names);
 void assertGatesExist(shared_ptr<Airport> airport, const vector<string>& names);
 void assertTaxiEdgesExist(shared_ptr<Airport> airport, const unordered_set<string>& names);
